+++ demoSpring/src/main/java/com/example/demo/repos/CustomerDao.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.example.demo.Entityexamples.Customer;
 @Repository
 @Transactional
-public interface CustomerDao extends JpaRepository<Customer,Integer>
+public interface CustomerDao extends JpaRepository<Customer,Long>
 {
 	
 }
